+++ packages/components/src/Form/Fields/FieldCollapseLabel.tsx
@@ -24,21 +24,13 @@
 
  */
 
-export * from './FieldCheckbox'
-export * from './FieldCheckboxGroup'
-export * from './FieldCollapseLabel'
-export * from './FieldColor'
-export * from './FieldDate'
-export * from './FieldDateRange'
-export * from './FieldRadio'
-export * from './FieldRadioGroup'
-export * from './FieldRangeSlider'
-export * from './FieldSelect'
-export * from './FieldSelectMulti'
-export * from './FieldSlider'
-export * from './FieldText'
-export * from './FieldTime'
-export * from './FieldTimeSelect'
-export * from './FieldTextArea'
-export * from './FieldToggleSwitch'
-export * from './Field'
+import styled from 'styled-components'
+import { Label } from '..'
+
+export const FieldCollapseLabel = styled.div`
+  ${Label} {
+    visibility: hidden;
+    height: 0;
+    padding: 0;
+  }
+`