+++ storybook/src/Forms/FieldCollapseLabel.stories.tsx
@@ -24,21 +24,26 @@
 
  */
 
-export * from './FieldCheckbox'
-export * from './FieldCheckboxGroup'
-export * from './FieldCollapseLabel'
-export * from './FieldColor'
-export * from './FieldDate'
-export * from './FieldDateRange'
-export * from './FieldRadio'
-export * from './FieldRadioGroup'
-export * from './FieldRangeSlider'
-export * from './FieldSelect'
-export * from './FieldSelectMulti'
-export * from './FieldSlider'
-export * from './FieldText'
-export * from './FieldTime'
-export * from './FieldTimeSelect'
-export * from './FieldTextArea'
-export * from './FieldToggleSwitch'
-export * from './Field'
+import { FieldCollapseLabel, Fieldset, FieldText } from '@looker/components'
+import React from 'react'
+
+export const FieldCollapse = () => (
+  <>
+    <FieldCollapseLabel>
+      <Fieldset inline legend="This is the Legend">
+        <FieldText label="First Label" />
+        <FieldText label="Second Label" />
+        <FieldText
+          label="Third Label"
+          validationMessage={{
+            message: 'This is an error',
+            type: 'error',
+          }}
+        />
+      </Fieldset>
+    </FieldCollapseLabel>
+  </>
+)
+export default {
+  title: 'Forms/FieldCollapseLabel',
+}